Add - Add new entities
